requirements.txt
# NGX Oracle Keeper - Python Dependencies
# Install with: pip install -r requirements.txt

# Blockchain interaction
web3==6.11.0
eth-account==0.9.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests for NGX data fetching
requests==2.31.0

# HTML parsing for web scraping
beautifulsoup4==4.12.2

# Task scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging
colorlog==6.8.0